+++ base_aux/aux_types/m1_type_aux.py
@@ -278,7 +278,7 @@ class TypeAux(Init_Source):
             """
             created specially for
             ---------------------
-            DictDotsAnnotRequired(dict)
+            DictAttrAnnotRequired(dict)
             it is not working without it!!!
             """
             mro = ()