+++ tests/aux_dicts/test__m2_dots_2_annots_required.py
@@ -1,7 +1,7 @@
 import pytest
 
 from base_aux.lambdas.m1_lambdas import *
-from base_aux.aux_dict.m3_dict_dots1_simple import *
+from base_aux.aux_dict.m3_dict_attr1_simple import *
 
 
 # =====================================================================================================================
@@ -11,16 +11,16 @@ dict_example = {
 }
 
 
-class Victim(DictDotsAnnotRequired):
+class Victim(DictAttrAnnotRequired):
     lowercase: str
 
 
 # =====================================================================================================================
 def test__obj():
-    # victim = DictDotsAnnotRequired()
+    # victim = DictAttrAnnotRequired()
     # assert victim == {}
     #
-    # victim = DictDotsAnnotRequired(hello=1)
+    # victim = DictAttrAnnotRequired(hello=1)
     # assert victim == {"hello": 1}
 
     try:
@@ -32,15 +32,15 @@ def test__obj():
 
 
 def test__dict_only():
-    assert LambdaTrySuccess(DictDotsAnnotRequired) == True
-    assert LambdaTrySuccess(DictDotsAnnotRequired)
+    assert LambdaTrySuccess(DictAttrAnnotRequired) == True
+    assert LambdaTrySuccess(DictAttrAnnotRequired)
 
-    assert LambdaTryFail(DictDotsAnnotRequired) != True
-    assert not LambdaTryFail(DictDotsAnnotRequired)
+    assert LambdaTryFail(DictAttrAnnotRequired) != True
+    assert not LambdaTryFail(DictAttrAnnotRequired)
 
-    assert LambdaTrySuccess(DictDotsAnnotRequired, **dict_example)
-    assert LambdaTrySuccess(DictDotsAnnotRequired, lowercase="lowercase")
-    assert LambdaTrySuccess(DictDotsAnnotRequired, LOWERCASE="lowercase")
+    assert LambdaTrySuccess(DictAttrAnnotRequired, **dict_example)
+    assert LambdaTrySuccess(DictAttrAnnotRequired, lowercase="lowercase")
+    assert LambdaTrySuccess(DictAttrAnnotRequired, LOWERCASE="lowercase")
 
 
 def test__with_annots():